+++ src/Sulu/Bundle/SecurityBundle/Controller/UserController.php
@@ -76,19 +76,9 @@ class UserController extends RestController implements ClassResourceInterface, S
             'email',
             $this->container->getParameter('sulu.model.user.class')
         );
-        $this->fieldDescriptors['password'] = new DoctrineFieldDescriptor(
-            'password',
-            'password',
-            $this->container->getParameter('sulu.model.user.class')
-        );
         $this->fieldDescriptors['locale'] = new DoctrineFieldDescriptor(
             'locale',
             'locale',
-            $this->container->getParameter('sulu.model.user.class')
-        );
-        $this->fieldDescriptors['salt'] = new DoctrineFieldDescriptor(
-            'salt',
-            'salt',
             $this->container->getParameter('sulu.model.user.class')
         );
         $this->fieldDescriptors['apiKey'] = new DoctrineFieldDescriptor(
@@ -269,7 +259,7 @@ class UserController extends RestController implements ClassResourceInterface, S
         if (null == $request->get('username')) {
             throw new MissingArgumentException($this->container->getParameter('sulu.model.user.class'), 'username');
         }
-        if (null === $request->get('password')) {
+        if ($request->isMethod('POST') && null === $request->get('password')) {
             throw new MissingArgumentException($this->container->getParameter('sulu.model.user.class'), 'password');
         }
         if (null == $request->get('locale')) {